requirements.txt
chris_plugin==0.1.0
pycivet==0.0.11
loguru~=0.6.0
